feeler/requirements.txt
django 
djangorestframework
tensorflow
numpy
pandas
psycopg2-binary
djangorestframework-simplejwt
python-dotenv
django-cors-headers
